# doc/api.yaml
openapi: 3.0.0
info: 
  title: WASA Photo
  description:  |
    APIs designed for a social focused on photos
  version: 1.0.0
...
paths:
...
  /users/{userId}/changeUsername:
    
    put:
      tags:
        - User
      summary: Change username
      description:  Change username of user of current session
      security:
        - bearerAuth: []
      operationId: setMyUserName
      parameters:
        - name: userId
          in: path
          description: Username of current user
          required: true
          schema:
            $ref: '#/components/schemas/Identifier'
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Username"
      responses:
        '200':
          description: Username changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /photos/{photoId}:
    
    parameters:
      - name: photoId
        in: path
        description: Identifier of photo showed
        required: true
        schema:
          $ref: "#/components/schemas/Identifier"
    
    get:
      tags:
        - Photo
      summary: Show photo
      description: Show a specific photo
      security:
        - bearerAuth: []
      operationId: showPhoto
      
      responses:
        '200':
          description: Photo opened successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Photo
      summary: Delete a photo
      description: Delete a specific photo of the current user
      security:
        - bearerAuth: []
      operationId: deletePhoto

      responses:
        '204': 
          description: Photo deleted successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /photos/{photoId}/comments/:

    parameters:
      - name: photoId
        in: path
        description: Identifier of the photo 
        required: true
        schema: 
          $ref: '#/components/schemas/Identifier'

    get:
      tags:
        - Comment
      summary: List comments
      description: List all comments of a photo
      security:
        - bearerAuth: []
      operationId: listComments
      responses: 
        '200':
          description: Comments listed successfully
          content:
            application/json:
              schema:
                description: List of comments
                type: array
                minItems: 0
                maxItems: 50
                items:
                  $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      
    post:
      tags:
        - Comment
      summary: Leave a comment
      description: Leave a comment for a specific photo
      security:
        - bearerAuth: []
      operationId: commentPhoto
      requestBody:
        content:
          text/plain:
            schema:
              type: string
              minLength: 1
              maxLength: 5000
              pattern: '^.*?$'
              description: A simple text
              example: Sample text
      responses:
        '201':
          description: Photo commented successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /photos/{photoId}/comments/{commentId}:

    parameters: 
      - name: photoId
        in: path
        description: Identifier of photo
        required: true
        schema:
          $ref: '#/components/schemas/Identifier'
      - name: commentId
        in: path
        description: Identifier of comment to modify
        required: true
        schema: 
          $ref: '#/components/schemas/Identifier'

    get:
      tags:
        - Comment
      summary: Get specific comment
      description: Return a specific comment of a specific photo
      security:
        - bearerAuth: []
      operationId: getComment
      responses:
        '200':
          description: Comment returned succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

#    put: 
#      tags:
#        - Comment
#      summary: Modify a comment
#      description: Modify a specific comment
#      security:
#        - bearerAuth: []
#      operationId: modifyComment
#      responses:
#        '200':
#          description: Comment modified successfully
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/ApiResponse'
#        '400':
#          description: Invalid request
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/ApiResponse'
#        '500':
#          description: The server encountered an internal error
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Comment
      summary: Delete a comment
      description: Delete a specific comment
      security:
        - bearerAuth: []
      operationId: uncommentPhoto

      responses:
        '204': 
          description: Comment deleted successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /photos/{photoId}/likes/:
    
    get:
      tags:
        - Like
      summary: List likes
      description: List all likes assigned to a photo
      security:
        - bearerAuth: []
      operationId: listLikes
      parameters:
        - name: photoId
          in: path
          description: Identifier of photo to like
          required: true
          schema:
            $ref: '#/components/schemas/Identifier'
      responses:
        '200':
          description: List of likes returned successfully
          content:
            application/json: 
              schema:
                description: List of likes
                type: array
                minItems: 0
                maxItems: 100
                items:
                  $ref: "#/components/schemas/Like"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /photos/{photoId}/likes/{userId}:

    parameters:
      - name: photoId
        in: path
        description: Identifier of photo to like
        required: true
        schema:
          $ref: '#/components/schemas/Identifier'
      - name: userId
        in: path
        description: Identifier of user who's leaving like
        required: true
        schema:
          $ref: '#/components/schemas/Identifier'

    post:
      tags:
        - Like
      summary: Leave a like
      description: Leave a like to a specific photo
      security:
        - bearerAuth: []
      operationId: likePhoto

      responses:
        '201':
          description: Photo liked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Like'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Like
      summary: Remove a like
      description: Remove like to a specific photo 
      security:
        - bearerAuth: []
      operationId: unlikePhoto
      
      responses:
        '204':
          description: Like removed successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/{userId}/following/:
  
    get:
      tags:
        - Following
      summary: List following
      description: Get list of users followed by a specific user
      parameters:
        - name: userId
          in: path
          description: Identifier of user who's following
          required: true
          schema:
            $ref: '#/components/schemas/Identifier'
      security:
        - bearerAuth: []
      operationId: listFollowed
      
      responses: 
        '200':
          description: List of following returned succesfully
          content:
            application/json:
              schema:
                description: List of user
                type: array
                minItems: 0
                maxItems: 40
                items:
                  $ref: '#/components/schemas/Identifier'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/{userId}/following/{userIdToFollow}:
  
    parameters:
      - name: userId
        in: path
        description: Identifier of user who's following
        required: true
        schema:
          $ref: '#/components/schemas/Identifier'
      - name: userIdToFollow
        in: path
        description: Identifier of user to follow
        required: true
        schema:
          $ref: '#/components/schemas/Identifier'
  
    put:
      tags:
        - Following
      summary: Follow an user
      description: Follow a specific user
      security:
        - bearerAuth: []
      operationId: followUser

      responses:
        '200':
          description: User followed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400': 
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Following
      summary: Unfollow an user
      description: Unfollow a specific user
      security:
        - bearerAuth: []
      operationId: unfollowUser
      
      responses:
        '204':
          description: User unfollowed successfully
        '400': 
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /users/{userId}/banned/{userIdToBan}:
  
    parameters:
      - name: userId
        in: path
        description: Identifier of user who's following
        required: true
        schema:
          $ref: '#/components/schemas/Identifier'
      - name: userIdToBan
        in: path
        description: Identifier of user to follow
        required: true
        schema:
          $ref: '#/components/schemas/Identifier'
  
    put:
      tags:
        - Ban
      summary: Ban an user
      description: Ban a specific user
      security:
        - bearerAuth: []
      operationId: banUser
      
      responses:
        '200':
          description: User banned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400': 
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Ban
      summary: Unban an user
      description: Unban a specific user
      security:
        - bearerAuth: []
      operationId: unbanUser
      
      responses:
        '204':
          description: User unbanned successfully
        '400': 
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: Bearer
  schemas:
    Identifier:
      readOnly: true
      description: Unique identifier used for several components
      type: integer
      example: 6365 

#    UserIdentifier:
#      description: Unique identifier for users 
#      type: string
#      pattern: "[a-z0-9]+"
#      minLength: 3
#      maxLength: 16
#      example: abcdef012345
    Username:
      description: Name of an user
      type: string
      pattern: '^.*?$'
      minLength: 4
      maxLength: 24 
      example: FraDepe 
    User:   #Utente
      required:
        - id
        - follower
        - following
        - username
      type: object
      description: User of platform
      properties:
        id:
          $ref: '#/components/schemas/Identifier'
        username:
          $ref: '#/components/schemas/Username'
        follower:
          type: array
          minItems: 0
          maxItems: 50
          description: List of people following this user
          items:
            $ref: '#/components/schemas/Identifier'
        following:
          type: array
          minItems: 0
          maxItems: 50
          description: List of people this user is following
          items:
            $ref: '#/components/schemas/Identifier'
        banned:
          type: array
          minItems: 0
          maxItems: 50
          description: List of user banned
          items:
            $ref: "#/components/schemas/Identifier"
        photos:
          type: array
          minItems: 0
          maxItems: 40
          description: List of photos
          items:
            $ref: "#/components/schemas/Photo"
    Photo:   
      required:
        - photo_id
        - photo_url
        - likes
        - date_time
        - comments
      type: object
      description: Photo of platform
      properties:
        photo_id:
          $ref: '#/components/schemas/Identifier'
        photo_file:
          description: File of the photo
          type: string
          format: binary
          minLength: 1
          maxLength: 99999
        likes:
          type: array
          minItems: 0
          maxItems: 50
          description: List of likes
          items:
            $ref: "#/components/schemas/Like"
        date_time:
          type: string
          minLength: 20
          maxLength: 20
          format: date-time
          description: Date and time of post 
          example: "2017-07-21T17:32:28Z"
        comments:
          type: array
          minItems: 0
          maxItems: 50
          description: List of comments
          items:
            $ref: "#/components/schemas/Comment"
    Comment:  #A comment
      required:
        - comment_id
        - text
        - userId
      type: object
      description: Comment to a photo
      properties:
        comment_id: 
          $ref: '#/components/schemas/Identifier'
        userId:
          $ref: '#/components/schemas/Identifier'
        text:
          type: string
          minLength: 1
          maxLength: 5000
          pattern: '^.*?$'
          description: A simple text
          example: Sample text
    Like:
      required:
        - photoId
        - user_Id
      type: object
      description: Like to a photo
      properties:
        photoId:
          $ref: '#/components/schemas/Identifier'
        user_Id:
          $ref: '#/components/schemas/Identifier'
    ApiResponse:
      type: object
      description: Text of responses
      properties:
        code:
          type: integer
          format: int32
          description: Response code
        type:
          type: string
          minLength: 1
          maxLength: 50
          description: Type of response
        message:
          type: string
          minLength: 1
          maxLength: 5000
          description: Text of response